+++ heartbeat/monitors/active/tcp/helpers_test.go
@@ -1,0 +1,80 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package tcp
+
+import (
+	"net"
+	"net/http"
+	"net/http/httptest"
+	"testing"
+	"time"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/elastic/beats/v7/heartbeat/hbtest"
+	"github.com/elastic/beats/v7/heartbeat/monitors/wrappers"
+	"github.com/elastic/beats/v7/heartbeat/scheduler/schedule"
+	"github.com/elastic/beats/v7/libbeat/beat"
+	"github.com/elastic/beats/v7/libbeat/common"
+)
+
+func testTCPConfigCheck(t *testing.T, configMap common.MapStr, host string, port uint16) *beat.Event {
+	config, err := common.NewConfigFrom(configMap)
+	require.NoError(t, err)
+
+	jobs, endpoints, err := create("tcp", config)
+	require.NoError(t, err)
+
+	sched, _ := schedule.Parse("@every 1s")
+	job := wrappers.WrapCommon(jobs, "test", "", "tcp", sched, time.Duration(0))[0]
+
+	event := &beat.Event{}
+	_, err = job(event)
+	require.NoError(t, err)
+
+	require.Equal(t, 1, endpoints)
+
+	return event
+}
+
+func setupServer(t *testing.T, serverCreator func(http.Handler) *httptest.Server) (*httptest.Server, uint16) {
+	server := serverCreator(hbtest.HelloWorldHandler(200))
+
+	port, err := hbtest.ServerPort(server)
+	require.NoError(t, err)
+
+	return server, port
+}
+
+// newLocalhostTestServer starts a server listening on the IP resolved from `localhost`
+// httptest.NewServer() binds explicitly on 127.0.0.1 (or [::1] if ipv4 is not available).
+// The IP resolved from `localhost` can be a different one, like 127.0.1.1.
+func newLocalhostTestServer(handler http.Handler) *httptest.Server {
+	listener, err := net.Listen("tcp", "localhost:0")
+	if err != nil {
+		panic("failed to listen on localhost: " + err.Error())
+	}
+
+	server := &httptest.Server{
+		Listener: listener,
+		Config:   &http.Server{Handler: handler},
+	}
+	server.Start()
+
+	return server
+}
