+++ heartbeat/monitors/active/tcp/tls_test.go
@@ -1,0 +1,162 @@
+// Licensed to Elasticsearch B.V. under one or more contributor
+// license agreements. See the NOTICE file distributed with
+// this work for additional information regarding copyright
+// ownership. Elasticsearch B.V. licenses this file to you under
+// the Apache License, Version 2.0 (the "License"); you may
+// not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package tcp
+
+import (
+	"crypto/x509"
+	"net"
+	"net/http/httptest"
+	"net/url"
+	"os"
+	"testing"
+	"time"
+
+	"github.com/elastic/beats/v7/heartbeat/monitors/wrappers"
+	"github.com/elastic/beats/v7/heartbeat/scheduler/schedule"
+	"github.com/elastic/beats/v7/libbeat/beat"
+	"github.com/elastic/beats/v7/libbeat/common"
+
+	"github.com/stretchr/testify/require"
+
+	"github.com/elastic/beats/v7/heartbeat/hbtest"
+	"github.com/elastic/beats/v7/heartbeat/monitors"
+	"github.com/elastic/go-lookslike"
+	"github.com/elastic/go-lookslike/testslike"
+)
+
+// Tests that we can check a TLS connection with a cert for a SAN IP
+func TestTLSSANIPConnection(t *testing.T) {
+	ip, port, cert, certFile, teardown := setupTLSTestServer(t)
+	defer teardown()
+
+	event := testTLSTCPCheck(t, ip, port, certFile.Name(), monitors.StdResolver)
+	testslike.Test(
+		t,
+		lookslike.Strict(lookslike.Compose(
+			hbtest.TLSChecks(0, 0, cert),
+			hbtest.RespondingTCPChecks(),
+			hbtest.BaseChecks(ip, "up", "tcp"),
+			hbtest.SummaryChecks(1, 0),
+			hbtest.SimpleURLChecks(t, "ssl", ip, port),
+		)),
+		event.Fields,
+	)
+}
+
+func TestTLSHostname(t *testing.T) {
+	ip, port, cert, certFile, teardown := setupTLSTestServer(t)
+	defer teardown()
+
+	hostname := cert.DNSNames[0] // Should be example.com
+	resolver := monitors.CreateStaticResolver()
+	resolver.Add(hostname, ip)
+	event := testTLSTCPCheck(t, hostname, port, certFile.Name(), resolver)
+	testslike.Test(
+		t,
+		lookslike.Strict(lookslike.Compose(
+			hbtest.TLSChecks(0, 0, cert),
+			hbtest.RespondingTCPChecks(),
+			hbtest.BaseChecks(ip, "up", "tcp"),
+			hbtest.SummaryChecks(1, 0),
+			hbtest.SimpleURLChecks(t, "ssl", hostname, port),
+			hbtest.ResolveChecks(ip),
+		)),
+		event.Fields,
+	)
+}
+
+func TestTLSInvalidCert(t *testing.T) {
+	ip, port, cert, certFile, teardown := setupTLSTestServer(t)
+	defer teardown()
+
+	mismatchedHostname := "notadomain.elastic.co"
+	resolver := monitors.CreateStaticResolver()
+	resolver.Add(cert.DNSNames[0], ip)
+	resolver.Add(mismatchedHostname, ip)
+	event := testTLSTCPCheck(t, mismatchedHostname, port, certFile.Name(), resolver)
+
+	testslike.Test(
+		t,
+		lookslike.Strict(lookslike.Compose(
+			hbtest.RespondingTCPChecks(),
+			hbtest.BaseChecks(ip, "down", "tcp"),
+			hbtest.SummaryChecks(0, 1),
+			hbtest.SimpleURLChecks(t, "ssl", mismatchedHostname, port),
+			hbtest.ResolveChecks(ip),
+			lookslike.MustCompile(map[string]interface{}{
+				"error": map[string]interface{}{
+					"message": x509.HostnameError{Certificate: cert, Host: mismatchedHostname}.Error(),
+					"type":    "io",
+				},
+			}),
+		)),
+		event.Fields,
+	)
+}
+
+func setupTLSTestServer(t *testing.T) (ip string, port uint16, cert *x509.Certificate, certFile *os.File, teardown func()) {
+	// Start up a TLS Server
+	server, port := setupServer(t, httptest.NewTLSServer)
+
+	// Parse its URL
+	serverURL, err := url.Parse(server.URL)
+	require.NoError(t, err)
+
+	// Determine the IP address the server's hostname resolves to
+	ips, err := net.LookupHost(serverURL.Hostname())
+	require.NoError(t, err)
+	require.Len(t, ips, 1)
+	ip = ips[0]
+
+	// Parse the cert so we can test against it
+	cert, err = x509.ParseCertificate(server.TLS.Certificates[0].Certificate[0])
+	require.NoError(t, err)
+
+	// Save the server's cert to a file so heartbeat can use it
+	certFile = hbtest.CertToTempFile(t, cert)
+	require.NoError(t, certFile.Close())
+
+	return ip, port, cert, certFile, func() {
+		os.Remove(certFile.Name())
+		server.Close()
+	}
+}
+
+func testTLSTCPCheck(t *testing.T, host string, port uint16, certFileName string, resolver monitors.Resolver) *beat.Event {
+	config, err := common.NewConfigFrom(common.MapStr{
+		"hosts":   host,
+		"ports":   int64(port),
+		"ssl":     common.MapStr{"certificate_authorities": certFileName},
+		"timeout": "1s",
+	})
+	require.NoError(t, err)
+
+	jobs, endpoints, err := createWithResolver(config, resolver)
+	require.NoError(t, err)
+
+	sched, _ := schedule.Parse("@every 1s")
+	job := wrappers.WrapCommon(jobs, "test", "", "tcp", sched, time.Duration(0))[0]
+
+	event := &beat.Event{}
+	_, err = job(event)
+	require.NoError(t, err)
+
+	require.Equal(t, 1, endpoints)
+
+	return event
+}
